# Bank-Application-FrontEnd-Springboot-angular/src/app/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/users/": {
      "get": {
        "tags": ["user"],
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserDto" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["user"],
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/transaction/": {
      "get": {
        "tags": ["transaction"],
        "operationId": "findAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TransactionDto" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["transaction"],
        "operationId": "save_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/contact/": {
      "get": {
        "tags": ["contact"],
        "operationId": "findAll_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ContactDto" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["contact"],
        "operationId": "save_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["authentication"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": ["authentication"],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthenticationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/address/": {
      "get": {
        "tags": ["address"],
        "operationId": "findAll_3",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AddressDto" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["address"],
        "operationId": "save_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddressDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/account/": {
      "get": {
        "tags": ["account"],
        "operationId": "findAll_4",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AccountDto" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["account"],
        "operationId": "save_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/users/validate/{user-id}": {
      "patch": {
        "tags": ["user"],
        "operationId": "validateAccount",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/users/invalidate/{user-id}": {
      "patch": {
        "tags": ["user"],
        "operationId": "invalidateAccount",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/users/{user-id}": {
      "get": {
        "tags": ["user"],
        "operationId": "findById",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/transaction/{transaction-id}": {
      "get": {
        "tags": ["transaction"],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "transaction-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionDto" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["transaction"],
        "operationId": "delete",
        "parameters": [
          {
            "name": "transaction-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/transaction/users/{user-id}": {
      "get": {
        "tags": ["transaction"],
        "operationId": "findAllByUserId",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TransactionDto" }
                }
              }
            }
          }
        }
      }
    },
    "/statistics/highest-transfert/{user-id}": {
      "get": {
        "tags": ["statistic"],
        "operationId": "highestTransfert",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "number" } }
            }
          }
        }
      }
    },
    "/statistics/highest-deposit/{user-id}": {
      "get": {
        "tags": ["statistic"],
        "operationId": "highestDeposit",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "number" } }
            }
          }
        }
      }
    },
    "/statistics/by-date/{user-id}": {
      "get": {
        "tags": ["statistic"],
        "operationId": "findSumTransactionByDate",
        "parameters": [
          {
            "name": "start-date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "end-date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionsSumDetails"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/statistics/balance/{user-id}": {
      "get": {
        "tags": ["statistic"],
        "operationId": "getAccountBalance",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "number" } }
            }
          }
        }
      }
    },
    "/contact/{contact-id}": {
      "get": {
        "tags": ["contact"],
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "contact-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContactDto" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["contact"],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "contact-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/contact/users/{user-id}": {
      "get": {
        "tags": ["contact"],
        "operationId": "findAllByUserId_1",
        "parameters": [
          {
            "name": "user-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ContactDto" }
                }
              }
            }
          }
        }
      }
    },
    "/address/{address-id}": {
      "get": {
        "tags": ["address"],
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "address-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddressDto" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["address"],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "address-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/account/{account-id}": {
      "get": {
        "tags": ["account"],
        "operationId": "findById_4",
        "parameters": [
          {
            "name": "account-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountDto" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["account"],
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "account-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserDto": {
        "required": ["email", "firstName", "lastName", "password"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "password": { "maxLength": 16, "minLength": 8, "type": "string" },
          "active": { "type": "boolean" }
        }
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "amount": { "maximum": 10000, "minimum": 10, "type": "number" },
          "destinationIban": { "type": "string" },
          "transactionDate": { "type": "string", "format": "date-time" },
          "type": { "type": "string", "enum": ["TRANSFERT", "DEPOSIT"] },
          "userId": { "type": "integer", "format": "int32" }
        }
      },
      "ContactDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "userId": { "type": "integer", "format": "int32" },
          "iban": { "type": "string" }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      },
      "AuthenticationRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "AddressDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "street": { "type": "string" },
          "county": { "type": "string" },
          "city": { "type": "string" },
          "zipCode": { "type": "integer", "format": "int32" },
          "houseNumber": { "type": "integer", "format": "int32" },
          "userId": { "type": "integer", "format": "int32" }
        }
      },
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "iban": { "type": "string" },
          "userDto": { "$ref": "#/components/schemas/UserDto" }
        }
      },
      "TransactionsSumDetails": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "transactionDate": { "type": "string", "format": "date-time" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}